+++ resources/views/menu.blade.php
@@ -1,3 +1,3 @@
-<a href="{{ route('plugin.nmsdashwidgets.index') }}">
+<a href="{{ route('plugin.settings', 'nmsdashwidgets') }}">
     <i class="fa fa-th-large fa-fw" aria-hidden="true"></i> {{ __('Dashboard Widget Bundle') }}
 </a>